Start on editor API
