+++ app/proto/admin.proto
@@ -54,15 +54,7 @@ service Admin {
     // it's sent to the user in an email
   }
 
-  rpc CreateCommunity(CreateCommunityReq) returns (org.couchers.api.communities.Community) {}
-
-  rpc UpdateCommunity(UpdateCommunityReq) returns (org.couchers.api.communities.Community) {}
-
   rpc GetChats(GetChatsReq) returns (GetChatsRes) {}
-
-  rpc ListEventCommunityInviteRequests(ListEventCommunityInviteRequestsReq) returns (ListEventCommunityInviteRequestsRes) {}
-
-  rpc DecideEventCommunityInviteRequest(DecideEventCommunityInviteRequestReq) returns (DecideEventCommunityInviteRequestRes) {}
 
   rpc DeleteEvent(DeleteEventReq) returns (google.protobuf.Empty) {}
 
@@ -85,8 +77,6 @@ service Admin {
   rpc CreateAccountDeletionLink(CreateAccountDeletionLinkReq) returns (CreateAccountDeletionLinkRes) {}
 
   rpc AccessStats(AccessStatsReq) returns (AccessStatsRes) {}
-
-  rpc SendBlogPostNotification(SendBlogPostNotificationReq) returns (google.protobuf.Empty) {}
 }
 
 message UserDetails {
@@ -255,22 +245,6 @@ message CreateApiKeyReq {
 
 message CreateApiKeyRes {}
 
-message CreateCommunityReq {
-  string name = 1;
-  string description = 3;
-  int64 parent_node_id = 4;
-  repeated int64 admin_ids = 5;
-  string geojson = 6;
-}
-
-message UpdateCommunityReq {
-  int64 community_id = 1;
-  string name = 2;
-  string description = 3;
-  int64 parent_node_id = 4;
-  string geojson = 5;
-}
-
 message GetChatsReq {
   // username, email, or user id
   string user = 1;
@@ -279,37 +253,6 @@ message GetChatsReq {
 message GetChatsRes {
   string response = 2;
 }
-
-message ListEventCommunityInviteRequestsReq {
-  uint32 page_size = 1;
-  string page_token = 2;
-}
-
-message EventCommunityInviteRequest {
-  int64 event_community_invite_request_id = 1;
-
-  int64 user_id = 2;
-  string event_url = 3;
-
-  // assuming they have notifs enabled
-  uint32 approx_users_to_notify = 4;
-  int64 community_id = 5;
-}
-
-message ListEventCommunityInviteRequestsRes {
-  repeated EventCommunityInviteRequest requests = 1;
-
-  string next_page_token = 2;
-}
-
-message DecideEventCommunityInviteRequestReq {
-  int64 event_community_invite_request_id = 1;
-
-  // whether to approve or deny it
-  bool approve = 2;
-}
-
-message DecideEventCommunityInviteRequestRes {}
 
 message DeleteEventReq {
   int64 event_id = 1;
@@ -414,10 +357,4 @@ message AccessStat {
 
 message AccessStatsRes {
   repeated AccessStat stats = 1;
-}
-
-message SendBlogPostNotificationReq {
-  string url = 1;
-  string title = 2;
-  string blurb = 3;
 }
