+++ app/proto/editor.proto
@@ -1,0 +1,75 @@
+syntax = "proto3";
+
+package org.couchers.editor;
+
+import "google/protobuf/empty.proto";
+
+import "annotations.proto";
+import "communities.proto";
+
+service Editor {
+  option (auth_level) = AUTH_LEVEL_EDITOR;
+
+  rpc CreateCommunity(CreateCommunityReq) returns (org.couchers.api.communities.Community) {}
+
+  rpc UpdateCommunity(UpdateCommunityReq) returns (org.couchers.api.communities.Community) {}
+
+  rpc ListEventCommunityInviteRequests(ListEventCommunityInviteRequestsReq) returns (ListEventCommunityInviteRequestsRes) {}
+
+  rpc DecideEventCommunityInviteRequest(DecideEventCommunityInviteRequestReq) returns (DecideEventCommunityInviteRequestRes) {}
+
+  rpc SendBlogPostNotification(SendBlogPostNotificationReq) returns (google.protobuf.Empty) {}
+}
+
+message CreateCommunityReq {
+  string name = 1;
+  string description = 3;
+  int64 parent_node_id = 4;
+  repeated int64 admin_ids = 5;
+  string geojson = 6;
+}
+
+message UpdateCommunityReq {
+  int64 community_id = 1;
+  string name = 2;
+  string description = 3;
+  int64 parent_node_id = 4;
+  string geojson = 5;
+}
+
+message ListEventCommunityInviteRequestsReq {
+  uint32 page_size = 1;
+  string page_token = 2;
+}
+
+message EventCommunityInviteRequest {
+  int64 event_community_invite_request_id = 1;
+
+  int64 user_id = 2;
+  string event_url = 3;
+
+  // assuming they have notifs enabled
+  uint32 approx_users_to_notify = 4;
+  int64 community_id = 5;
+}
+
+message ListEventCommunityInviteRequestsRes {
+  repeated EventCommunityInviteRequest requests = 1;
+
+  string next_page_token = 2;
+}
+
+message DecideEventCommunityInviteRequestReq {
+  int64 event_community_invite_request_id = 1;
+
+  // whether to approve or deny it
+  bool approve = 2;
+}
+
+message DecideEventCommunityInviteRequestRes {}
+
+message SendBlogPostNotificationReq {
+  string url = 1;
+  string title = 2;
+  string blurb = 3;
+}
